services/georesolver/requirements.txt
-i https://pypi.org/simple
google-api-core==1.3.0
grpcio==1.14.2
protobuf==3.6.1
geoip2==2.9.0
